* 📝 docs(nvim/init.lua): add notes on leader key and fix typo Added notes on the leader key in Neovim and how to access its help documentation. Also fixed a typo in the Telescope command where "hepl_tags" was corrected to "help_tags".
* 🚀 chore(init.lua): remove comments and unused code
This commit removes comments and unused code from the init.lua file. The comments removed were introductory comments that were no longer necessary, and the code removed was related to the installation of a package manager that is no longer used.

* 🎨 style(nvim/init.lua): remove comments and unused language servers
* 🚀 chore(nvim/init.lua): add rust_analyzer language server
This commit removes comments that are not necessary and removes unused language servers. The rust_analyzer language server is added to the list of language servers.